+++ src/app/Components/REGISTER_YOUR_SAMPLES_HERE.cs
@@ -7,7 +7,9 @@ using Components.ComponentsSamples.Pickers;
 using Components.ComponentsSamples.Searching;
 using Components.Resources.LocalizedStrings;
 using Components.ResourcesSamples.Colors;
+using Components.ResourcesSamples.Icons;
 using Components.ResourcesSamples.Sizes;
+using DIPS.Mobile.UI.Resources.Icons;
 
 namespace Components;
 
@@ -26,8 +28,8 @@ public static class REGISTER_YOUR_SAMPLES_HERE
             new(SampleType.Resources, LocalizedStrings.Colors, () => new ColorsSamples()),
             new(SampleType.Resources, LocalizedStrings.Sizes, () => new SizesSamples()),
             new(SampleType.Components, LocalizedStrings.ListItems, () => new ListItemsSamples()),
-            new(SampleType.Components, LocalizedStrings.Chips, () => new ChipsSamples())
-
+            new(SampleType.Components, LocalizedStrings.Chips, () => new ChipsSamples()),
+            new(SampleType.Resources, LocalizedStrings.Icons, () => new IconsSamples())
         };
     }
 }