+++ src/library/DIPS.Mobile.UI/AppHostBuilderExtensions.cs
@@ -3,22 +3,9 @@ using DIPS.Mobile.UI.Components.ContextMenus;
 using DIPS.Mobile.UI.Components.Images;
 using DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker;
 using DIPS.Mobile.UI.Components.Searching;
-using DIPS.Mobile.UI.Effects.AwesomeTouchEffect;
-#if __ANDROID__
-using DIPS.Mobile.UI.Components.Buttons.Android;
-using DIPS.Mobile.UI.Components.Searching.Android;
-#elif __IOS__
-using DIPS.Mobile.UI.Components.Searching.iOS;
-#endif
-
 using Microsoft.Maui.LifecycleEvents;
 using ContextMenuPlatformEffect = DIPS.Mobile.UI.Components.ContextMenus.ContextMenuPlatformEffect;
-using DateAndTimePickerHandler = DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker.DateAndTimePickerHandler;
-using DatePicker = DIPS.Mobile.UI.Components.Pickers.DatePicker.DatePicker;
-using DatePickerHandler = DIPS.Mobile.UI.Components.Pickers.DatePicker.DatePickerHandler;
-using Image = DIPS.Mobile.UI.Components.Images.Image;
-using TimePicker = Microsoft.Maui.Controls.TimePicker;
-using TimePickerHandler = DIPS.Mobile.UI.Components.Pickers.TimePicker.TimePickerHandler;
+using SearchBar = DIPS.Mobile.UI.Components.Searching.SearchBar;
 
 namespace DIPS.Mobile.UI;
 
@@ -53,12 +40,13 @@ public static class AppHostBuilderExtensions
             handlers.AddHandler(typeof(DatePicker), typeof(DatePickerHandler));
             handlers.AddHandler(typeof(DateAndTimePicker), typeof(DateAndTimePickerHandler));
             handlers.AddHandler(typeof(Components.Pickers.TimePicker.TimePicker), typeof(TimePickerHandler));
+            handlers.AddHandler(typeof(NativeIcon), typeof(NativeIconHandler));
+            handlers.AddHandler(typeof(SearchBar), typeof(SearchBarHandler));
 #if __ANDROID__
-            handlers.AddHandler(typeof(InternalSearchBar), typeof(InternalSearchBarHandler));
-            handlers.AddHandler(typeof(IndeterminateProgressBar), typeof(IndeterminateProgressBarHandler));
-            handlers.AddHandler(typeof(Button), typeof(ButtonHandler));
+            handlers.AddHandler(typeof(Button), typeof(DIPS.Mobile.UI.Components.Buttons.Android.ButtonHandler));
+            handlers.AddHandler(typeof(DIPS.Mobile.UI.Components.Searching.Android.IndeterminateProgressBar), typeof(DIPS.Mobile.UI.Components.Searching.Android.IndeterminateProgressBarHandler));
 #elif __IOS__
-            handlers.AddHandler(typeof(InternalSearchBar), typeof(InternalSearchBarHandler));
+            handlers.AddHandler(typeof(DIPS.Mobile.UI.Components.Searching.iOS.InternalSearchBar), typeof(DIPS.Mobile.UI.Components.Searching.iOS.InternalSearchBarHandler));
 #endif
         });