+++ src/library/DIPS.Mobile.UI/Components/Images/NativeIcon.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.Components.Images
+{
+    public partial class NativeIcon : Image
+    {
+    }
+}