+++ src/library/DIPS.Mobile.UI/Components/Searching/SearchBar.Properties.cs
@@ -4,24 +4,14 @@ namespace DIPS.Mobile.UI.Components.Searching
 {
     public partial class SearchBar
     {
-        public static readonly BindableProperty IconsColorProperty = BindableProperty.Create(
-            nameof(IconsColor),
-            typeof(Color),
-            typeof(SearchBar));
-
         /// <summary>
         /// Sets the color of the icons that people see in the search bar.
         /// </summary>
-        public Color IconsColor
+        public Color? IconsColor
         {
             get => (Color)GetValue(IconsColorProperty);
             set => SetValue(IconsColorProperty, value);
         }
-
-        public static readonly BindableProperty IsBusyProperty = BindableProperty.Create(
-            nameof(IsBusy),
-            typeof(bool),
-            typeof(SearchBar));
 
         /// <summary>
         /// Displays a busy indication for people to see while they wait for the search bars results.
@@ -32,15 +22,23 @@ namespace DIPS.Mobile.UI.Components.Searching
             set => SetValue(IsBusyProperty, value);
         }
 
-        public static readonly BindableProperty ShowsCancelButtonProperty = BindableProperty.Create(
-            nameof(HasCancelButton),
-            typeof(bool),
-            typeof(SearchBar));
+        /// <summary>
+        /// The background color of the indeterminate busy indicator below the search bar that is visible when <see cref="IsBusy"/> is set to true.
+        /// </summary>
+        public Color? AndroidBusyBackgroundColor
+        {
+            get => (Color)GetValue(AndroidBusyBackgroundColorProperty);
+            set => SetValue(AndroidBusyBackgroundColorProperty, value);
+        }
 
-        public static readonly BindableProperty HasBusyIndicationProperty = BindableProperty.Create(
-            nameof(HasBusyIndication),
-            typeof(bool),
-            typeof(SearchBar));
+        /// <summary>
+        /// The color of the indeterminate busy indicator below the search bar that is visible when <see cref="IsBusy"/> is set to true.
+        /// </summary>
+        public Color? AndroidBusyColor
+        {
+            get => (Color)GetValue(AndroidBusyColorProperty);
+            set => SetValue(AndroidBusyColorProperty, value);
+        }
 
         /// <summary>
         /// Indicates that the busy indication should be visible for people when you use <see cref="IsBusy"/> property to indicate that the search bar is busy.
@@ -56,14 +54,9 @@ namespace DIPS.Mobile.UI.Components.Searching
         /// </summary>
         public bool HasCancelButton
         {
-            get => (bool)GetValue(ShowsCancelButtonProperty);
-            set => SetValue(ShowsCancelButtonProperty, value);
+            get => (bool)GetValue(HasCancelButtonProperty);
+            set => SetValue(HasCancelButtonProperty, value);
         }
-
-        public static readonly BindableProperty CancelCommandProperty = BindableProperty.Create(
-            nameof(CancelCommand),
-            typeof(ICommand),
-            typeof(SearchBar));
 
         /// <summary>
         /// The command to be executed when people tap the cancel button.
@@ -74,11 +67,6 @@ namespace DIPS.Mobile.UI.Components.Searching
             set => SetValue(CancelCommandProperty, value);
         }
 
-        public static readonly BindableProperty CancelCommandParameterProperty = BindableProperty.Create(
-            nameof(CancelCommandParameter),
-            typeof(object),
-            typeof(SearchBar));
-
         /// <summary>
         /// The parameter to be sent with the <see cref="CancelCommand"/> when people tap the cancel button.
         /// </summary>
@@ -87,11 +75,6 @@ namespace DIPS.Mobile.UI.Components.Searching
             get => (object)GetValue(CancelCommandParameterProperty);
             set => SetValue(CancelCommandParameterProperty, value);
         }
-
-        public static readonly BindableProperty PlaceholderProperty = BindableProperty.Create(
-            nameof(Placeholder),
-            typeof(string),
-            typeof(SearchBar));
 
         /// <summary>
         /// The placeholder for people to see before searching.
@@ -102,17 +85,6 @@ namespace DIPS.Mobile.UI.Components.Searching
             set => SetValue(PlaceholderProperty, value);
         }
 
-        public static readonly BindableProperty TextProperty = BindableProperty.Create(
-            nameof(Text),
-            typeof(string),
-            typeof(SearchBar), propertyChanged: (bindable, value, newValue) =>
-            {
-                if (bindable is SearchBar searchBar && value is string oldTextValue && newValue is string newTextValue)
-                {
-                    searchBar.TextChanged?.Invoke(searchBar, new TextChangedEventArgs(oldTextValue, newTextValue));
-                }
-            });
-
         /// <summary>
         /// The text that people search for when using the search bar.
         /// </summary>
@@ -121,11 +93,6 @@ namespace DIPS.Mobile.UI.Components.Searching
             get => (string)GetValue(TextProperty);
             set => SetValue(TextProperty, value);
         }
-
-        public static readonly BindableProperty TextColorProperty = BindableProperty.Create(
-            nameof(TextColor),
-            typeof(Color),
-            typeof(SearchBar));
 
         /// <summary>
         /// The color of the text for people to see.
@@ -136,29 +103,19 @@ namespace DIPS.Mobile.UI.Components.Searching
             set => SetValue(TextColorProperty, value);
         }
 
-        public static readonly BindableProperty CancelButtonColorProperty = BindableProperty.Create(
-            nameof(CancelButtonColor),
-            typeof(Color),
-            typeof(SearchBar));
-
         /// <summary>
-        /// The color of the text in the cancel button to see.
+        /// The color of the text in the cancel button.
         /// </summary>
-        public Color CancelButtonColor
+        public Color CancelButtonTextColor
         {
-            get => (Color)GetValue(CancelButtonColorProperty);
-            set => SetValue(CancelButtonColorProperty, value);
+            get => (Color)GetValue(CancelButtonTextColorProperty);
+            set => SetValue(CancelButtonTextColorProperty, value);
         }
 
         /// <summary>
         /// Event that gets invoked when people search using the search bar.
         /// </summary>
         public event EventHandler<TextChangedEventArgs> TextChanged;
-
-        public static readonly BindableProperty SearchCommandProperty = BindableProperty.Create(
-            nameof(SearchCommand),
-            typeof(ICommand),
-            typeof(SearchBar));
 
         /// <summary>
         /// The search command that gets invoked when people has tapped the search button in the keyboard.
@@ -168,44 +125,113 @@ namespace DIPS.Mobile.UI.Components.Searching
             get => (ICommand)GetValue(SearchCommandProperty);
             set => SetValue(SearchCommandProperty, value);
         }
-        
-        public static readonly BindableProperty TextFieldColorProperty = BindableProperty.Create(
-            nameof(TextFieldColor),
-            typeof(Color),
-            typeof(SearchBar));
 
         /// <summary>
         /// The background color of the textfield.
         /// </summary>
-        public Color TextFieldColor
+        public Color? iOSSearchFieldBackgroundColor
         {
-            get => (Color)GetValue(TextFieldColorProperty);
-            set => SetValue(TextFieldColorProperty, value);
+            get => (Color)GetValue(iOSSearchFieldBackgroundColorProperty);
+            set => SetValue(iOSSearchFieldBackgroundColorProperty, value);
         }
-
-        public static readonly BindableProperty BarColorProperty = BindableProperty.Create(
-            nameof(BarColor),
-            typeof(Color),
-            typeof(SearchBar));
 
         /// <summary>
         /// The background color of the search bar.
         /// </summary>
-        public Color BarColor
+        public Color? BarColor
         {
             get => (Color)GetValue(BarColorProperty);
             set => SetValue(BarColorProperty, value);
         }
 
+        /// <summary>
+        /// The color of the place holder in the search bar.
+        /// </summary>
+        public Color? PlaceholderColor
+        {
+            get => (Color)GetValue(PlaceholderColorProperty);
+            set => SetValue(PlaceholderColorProperty, value);
+        }
+        
+           public static readonly BindableProperty IconsColorProperty = BindableProperty.Create(
+            nameof(IconsColor),
+            typeof(Color),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty HasCancelButtonProperty = BindableProperty.Create(
+            nameof(HasCancelButton),
+            typeof(bool),
+            typeof(SearchBar));
+
+        public static readonly BindableProperty HasBusyIndicationProperty = BindableProperty.Create(
+            nameof(HasBusyIndication),
+            typeof(bool),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty IsBusyProperty = BindableProperty.Create(
+            nameof(IsBusy),
+            typeof(bool),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty AndroidBusyBackgroundColorProperty = BindableProperty.Create(
+            nameof(AndroidBusyBackgroundColor),
+            typeof(Color),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty AndroidBusyColorProperty = BindableProperty.Create(
+            nameof(AndroidBusyColor),
+            typeof(Color),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty CancelCommandProperty = BindableProperty.Create(
+            nameof(CancelCommand),
+            typeof(ICommand),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty CancelCommandParameterProperty = BindableProperty.Create(
+            nameof(CancelCommandParameter),
+            typeof(object),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty TextProperty = BindableProperty.Create(
+            nameof(Text),
+            typeof(string),
+            typeof(SearchBar), propertyChanged: OnTextChanged);
+        
+        public static readonly BindableProperty PlaceholderProperty = BindableProperty.Create(
+            nameof(Placeholder),
+            typeof(string),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty TextColorProperty = BindableProperty.Create(
+            nameof(TextColor),
+            typeof(Color),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty SearchCommandProperty = BindableProperty.Create(
+            nameof(SearchCommand),
+            typeof(ICommand),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty iOSSearchFieldBackgroundColorProperty = BindableProperty.Create(
+            nameof(iOSSearchFieldBackgroundColor),
+            typeof(Color),
+            typeof(SearchBar));
+        
         public static readonly BindableProperty PlaceholderColorProperty = BindableProperty.Create(
             nameof(PlaceholderColor),
             typeof(Color),
             typeof(SearchBar));
 
-        public Color PlaceholderColor
-        {
-            get => (Color)GetValue(PlaceholderColorProperty);
-            set => SetValue(PlaceholderColorProperty, value);
-        }
+        public static readonly BindableProperty CancelButtonTextColorProperty = BindableProperty.Create(
+            nameof(CancelButtonTextColor),
+            typeof(Color),
+            typeof(SearchBar));
+        
+        public static readonly BindableProperty BarColorProperty = BindableProperty.Create(
+            nameof(BarColor),
+            typeof(Color),
+            typeof(SearchBar));
+
     }
 }