+++ CHANGELOG.md
@@ -7,6 +7,16 @@
 - Added ListItems
 - Added TouchEffect
 
+## [3.0.1]
+- Minor changes to the search bar in order for it to work more smoothly.
+
+## [3.0.0]
+- Breaking change: Renamed dui:Image to dui:NativeIcon and simplified the ios and android properties for setting the icon.
+- Added <MauiImage> for our icons. This can be disabled by setting DIPSIncludeIcons = False in your .csproj.
+
+## [2.1.1]
+- SearchBar on Android is now showing if you have set HasCancelButton = false.
+
 ## [2.1.0]
 - Added DatePicker
 - Added ItemPicker